Work on adding function definitions.
Functions are meant to allow recursive queries in a simple natural
manner. Currently it is possible to write a recursive query if you
add pyflwor.compile into the namespace. Then with some syntactical hacks
you can construct new namespaces and write "subqueries" which can be
excuted with the constructed namespaces. It is clunky and not very
powerful. This new syntax should clean it up.